guidance and control stable
