+++ include/simulation/control/pid/controllers/velocity.hpp
@@ -9,12 +9,6 @@ namespace control { struct ControlOutput; struct VelocityControlLawInput; } // f
 
 namespace control {
 
-    struct VelocityPIDInput : VelocityControlLawInput {
-        dynamics::RigidBodyState& zN_t;
-        actuators::PropulsorActuators& propulsor_actuators;
-        guidance::VelocitySetpoint setpoint;
-    };
-
     struct VelocityPIDParameters : PIDControlLawParameters, VelocityControlLawParameters {};
 
     struct VelocityPID {
@@ -22,11 +16,9 @@ namespace control {
         PIDControlLaw policy;
 
         VelocityPID(const VelocityPIDParameters& params);
-        ControlOutput step(const VelocityPIDInput& ctrl_law_input);
+        ControlOutput step(const VelocityControlLawInput& ctrl_law_input);
 
-        PIDControlLawInput make_pid_control_law_input(
-            const VelocityPIDInput& ctrl_law_input
-        );
+        PIDControlLawInput make_pid_control_law_input(const VelocityControlLawInput& ctrl_law_input);
         std::tuple<double, double, double> allocate_thrust(double T_tot, const actuators::PropulsorActuators& propulsor_actuators);
     };
 }
